Jijesh | customer account service and endpoint to update existing customer

+++ dloserver/grails-app/domain/com/dlohaiti/dloserver/CustomerAccount.groovy
@@ -5,7 +5,7 @@ class CustomerAccount {
     String phoneNumber
     String address
     String contactName
-    Integer dueAmount = 0
+    Double dueAmount = 0
 
     CustomerType customerType
     Kiosk kiosk
